refactor: restructure unsubscribe logic into a dedicated UnsubscribeForm component
