--- app/unsubscribe/page.js
@@ -1,137 +1,0 @@
-'use client';
-"use strict";
-var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
-    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
-    return new (P || (P = Promise))(function (resolve, reject) {
-        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
-        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
-        step((generator = generator.apply(thisArg, _arguments || [])).next());
-    });
-};
-var __generator = (this && this.__generator) || function (thisArg, body) {
-    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g = Object.create((typeof Iterator === "function" ? Iterator : Object).prototype);
-    return g.next = verb(0), g["throw"] = verb(1), g["return"] = verb(2), typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
-    function verb(n) { return function (v) { return step([n, v]); }; }
-    function step(op) {
-        if (f) throw new TypeError("Generator is already executing.");
-        while (g && (g = 0, op[0] && (_ = 0)), _) try {
-            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
-            if (y = 0, t) op = [op[0] & 2, t.value];
-            switch (op[0]) {
-                case 0: case 1: t = op; break;
-                case 4: _.label++; return { value: op[1], done: false };
-                case 5: _.label++; y = op[1]; op = [0]; continue;
-                case 7: op = _.ops.pop(); _.trys.pop(); continue;
-                default:
-                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
-                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
-                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
-                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
-                    if (t[2]) _.ops.pop();
-                    _.trys.pop(); continue;
-            }
-            op = body.call(thisArg, _);
-        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
-        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
-    }
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.default = UnsubscribePage;
-var react_1 = require("react");
-var Button_1 = require("@/components/ui/Button");
-var navigation_1 = require("next/navigation");
-function UnsubscribePage() {
-    var _this = this;
-    var _a = (0, react_1.useState)(''), email = _a[0], setEmail = _a[1];
-    var _b = (0, react_1.useState)('idle'), status = _b[0], setStatus = _b[1];
-    var _c = (0, react_1.useState)(''), message = _c[0], setMessage = _c[1];
-    var searchParams = (0, navigation_1.useSearchParams)();
-    (0, react_1.useEffect)(function () {
-        var emailParam = searchParams.get('email');
-        if (emailParam) {
-            setEmail(emailParam);
-            if (status === 'idle') {
-                handleUnsubscribe(emailParam);
-            }
-        }
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, []); // <-- Only run on mount
-    var handleUnsubscribe = function (targetEmail) { return __awaiter(_this, void 0, void 0, function () {
-        var res, data;
-        return __generator(this, function (_a) {
-            switch (_a.label) {
-                case 0:
-                    setStatus('loading');
-                    return [4 /*yield*/, fetch('/api/unsubscribe', {
-                            method: 'POST',
-                            headers: { 'Content-Type': 'application/json' },
-                            body: JSON.stringify({ email: targetEmail || email }),
-                        })];
-                case 1:
-                    res = _a.sent();
-                    if (!res.ok) return [3 /*break*/, 2];
-                    setStatus('success');
-                    setMessage('You have been unsubscribed from our newsletter.');
-                    return [3 /*break*/, 4];
-                case 2:
-                    setStatus('error');
-                    return [4 /*yield*/, res.json()];
-                case 3:
-                    data = _a.sent();
-                    setMessage(data.error || 'Something went wrong.');
-                    _a.label = 4;
-                case 4: return [2 /*return*/];
-            }
-        });
-    }); };
-    var handleResubscribe = function () { return __awaiter(_this, void 0, void 0, function () {
-        var res, data;
-        return __generator(this, function (_a) {
-            switch (_a.label) {
-                case 0:
-                    setStatus('loading');
-                    return [4 /*yield*/, fetch('/api/resubscribe', {
-                            method: 'POST',
-                            headers: { 'Content-Type': 'application/json' },
-                            body: JSON.stringify({ email: email }),
-                        })];
-                case 1:
-                    res = _a.sent();
-                    if (!res.ok) return [3 /*break*/, 2];
-                    setStatus('resubscribed');
-                    setMessage('You have been resubscribed to our newsletter!');
-                    return [3 /*break*/, 4];
-                case 2:
-                    setStatus('error');
-                    return [4 /*yield*/, res.json()];
-                case 3:
-                    data = _a.sent();
-                    setMessage(data.error || 'Something went wrong.');
-                    _a.label = 4;
-                case 4: return [2 /*return*/];
-            }
-        });
-    }); };
-    var handleSubmit = function (e) { return __awaiter(_this, void 0, void 0, function () {
-        return __generator(this, function (_a) {
-            e.preventDefault();
-            handleUnsubscribe();
-            return [2 /*return*/];
-        });
-    }); };
-    var emailParam = searchParams.get('email');
-    return (<div className="max-w-md mx-auto mt-20 p-8 bg-white rounded-lg shadow">
-      <h2 className="text-2xl font-bold mb-4">Unsubscribe from Newsletter</h2>
-      <form onSubmit={handleSubmit} className="flex flex-col gap-4">
-        <input type="email" placeholder="Your email" className="w-full p-3 border rounded" value={email} onChange={function (e) { return setEmail(e.target.value); }} required disabled={status === 'loading' || !!emailParam}/>
-        <Button_1.default text={status === 'loading' ? 'Unsubscribing...' : 'Unsubscribe'} type="submit" variant="primary" className="w-full" disabled={status === 'loading' || !!emailParam}/>
-      </form>
-      {status === 'success' && (<div className="mt-6">
-          <p className="text-green-600 mb-4">{message}</p>
-          <Button_1.default text="Resubscribe" variant="secondary" className="w-full" onClick={handleResubscribe}/>
-        </div>)}
-      {status === 'resubscribed' && (<p className="text-green-600 mt-4">{message}</p>)}
-      {status === 'error' && <p className="text-red-600 mt-4">{message}</p>}
-    </div>);
-}
